added replacement icons and a Stylesheet for a darker ui, will be adding in a light/dark mode switch
