+++ budget_tool/ui.py
@@ -9,7 +9,7 @@ from functools import partial
 import platform
 from pathlib import Path
 
-# from utils import CollapsibleWidget
+from custom_widgets import CollapsibleWidget
 
 # QApplication setup
 QApplication = QtWidgets.QApplication
@@ -24,197 +24,8 @@ CURRENT_DIR = 'C:/Users/Kent/Desktop/budget_tool'
 
 # Build the path to the icons folder
 ICONS_DIR = os.path.join(CURRENT_DIR, "icons")
-
-def get_main_window():
-    """Return a handle to the Maya main window."""
-    return next(w for w in app.topLevelWidgets() if w.objectName() == "MayaWindow")
-
-
 # Class for setting up Widget under the Head in Collapsible Header
 # -------------------------------------------------------------------------------------------
-class CollapsibleHeader(QtWidgets.QWidget):
-    # Grabbing arrow pngs' from mayas' default UI library
-
-    COLLAPSED_PIXMAP = QtGui.QPixmap(":teRightArrow.png")
-    EXPANDED_PIXMAP = QtGui.QPixmap(":teDownArrow.png")
-
-    # Creating a signal for clicking the header
-
-    clicked = QtCore.pyqtSignal()
-
-    # function that auto-starts up when CollapsibleHeader() is called
-    def __init__(self, text, parent=None):
-        super(
-            CollapsibleHeader,
-            self,
-        ).__init__(parent)
-
-        # need to set this or the header won't get color despite setting
-        # pallete
-        self.setAutoFillBackground(True)
-        self.set_background_color(None)
-        # icon that uses the arrow pngs' and setting width to width of png
-        self.icon_label = QtWidgets.QLabel()
-        self.icon_label.setFixedWidth(self.COLLAPSED_PIXMAP.width())
-
-        # text label that uses text passes in function creation and enabling
-        # mouse clicked on it
-        self.text_label = QtWidgets.QLabel()
-        self.text_label.setAttribute(QtCore.Qt.WA_TransparentForMouseEvents)
-
-        # creating main layout and setting its margins, then adding all
-        # previous widgets to the layout
-        self.main_layout = QtWidgets.QHBoxLayout(self)
-        self.main_layout.setContentsMargins(
-            4,
-            4,
-            4,
-            4,
-        )
-
-
-        self.main_layout.addWidget(self.text_label)
-        self.main_layout.addWidget(self.icon_label)
-        self.main_layout.addWidget(self.text_label)
-
-        # setting defaults, calling below functions
-        self.set_text(text)
-        self.set_expanded(True)
-        self.setSizePolicy(
-            QtWidgets.QSizePolicy.Expanding,
-            QtWidgets.QSizePolicy.Fixed,
-        )
-
-    # sets text label with passed text and formats it as bold
-    def set_text(
-        self,
-        text,
-    ):
-        self.text_label.setText("<b>{0}</b>".format(text))
-
-    # gets the background color from qpushbutton pallete which carries mayas
-    # default grey
-    def set_background_color(self, color=None):
-        if color is None:
-            color = QtWidgets.QPushButton().palette().color(QtGui.QPalette.Button)
-        elif isinstance(color, tuple) and len(color) == 3:
-            color = QtGui.QColor(*color)
-
-        palette = self.palette()
-        palette.setColor(
-            QtGui.QPalette.Window,
-            color,
-        )
-        self.setPalette(palette)
-
-    # returns or = if widget is expanded
-    def is_expanded(
-        self,
-    ):
-        return self._expanded
-
-    # function that sets state of widget collapsed or expanded
-    def set_expanded(
-        self,
-        expanded,
-    ):
-        self._expanded = expanded
-
-        # changes the png arrow for either state
-        if self._expanded:
-            self.icon_label.setPixmap(self.EXPANDED_PIXMAP)
-        else:
-            self.icon_label.setPixmap(self.COLLAPSED_PIXMAP)
-
-    # overrides mouseReleaseEvent to emit a signal
-    def mouseReleaseEvent(
-        self,
-        event,
-    ):
-        self.clicked.emit()  # pylint: disable=E1101
-
-
-class CollapsibleWidget(QtWidgets.QWidget):
-    # function that auto-starts up when CollapsibleHeader() is called
-    def __init__(self, text, parent=None, window=None):
-        super(
-            CollapsibleWidget,
-            self,
-        ).__init__(parent)
-
-        self.tool_window = window
-        # calls for Header
-        self.header_wdg = CollapsibleHeader(text)
-        self.header_wdg.clicked.connect(self.on_header_clicked)
-
-        # Creates body widget for part below header
-        self.body_wdg = QtWidgets.QWidget()
-
-        # parents body layout to the body widget and sets spacing settings
-        self.body_layout = QtWidgets.QVBoxLayout(self.body_wdg)
-        self.body_layout.setContentsMargins(
-            4,
-            2,
-            4,
-            2,
-        )
-        self.body_layout.setSpacing(3)
-
-        # creates main layout to house header and the body widget
-        self.details_layout = QtWidgets.QVBoxLayout(self)
-        self.details_layout.setSpacing(1)
-        self.details_layout.setContentsMargins(
-            0,
-            0,
-            0,
-            0,
-        )
-        self.details_layout.addWidget(self.header_wdg)
-        self.details_layout.addWidget(self.body_wdg)
-
-        # set default states of collapsible widget
-        self.set_expanded(True)
-
-        # making addWidget function for class
-
-    def add_widget(
-        self,
-        widget,
-    ):
-        self.body_layout.addWidget(widget)
-
-        # making addLayout function for class
-
-    def add_layout(
-        self,
-        layout,
-    ):
-        self.body_layout.addLayout(layout)
-
-        # making setState function for class
-
-    def set_expanded(
-        self,
-        expanded,
-    ):
-        self.header_wdg.set_expanded(expanded)
-        self.body_wdg.setVisible(expanded)
-
-        # making easy color change function to class
-
-    def set_header_background_color(
-        self,
-        color,
-    ):
-        self.header_wdg.set_background_color(color)
-
-        # sends signal
-
-    def on_header_clicked(
-        self,
-    ):
-        self.set_expanded(not self.header_wdg.is_expanded())
-
 class BudgetingTool(QtWidgets.QDialog):
     """Main dialog for the Couples Budgeting Tool."""
 
@@ -222,14 +33,7 @@ class BudgetingTool(QtWidgets.QDialog):
         super(BudgetingTool, self).__init__(parent)
 
         self.setWindowTitle("Couples Budgeting Tool")
-        '''
-        self.setStyleSheet("""
-            QDialog {
-                background-color: #f0f0f0;
-                color: #202020;
-            }
-        """)
-        '''
+        apply_maya_style(self)
         self.open_on_details = open_on_details
         self.user_details = None 
         self.user_ui_elements = {}
@@ -1197,6 +1001,68 @@ class BudgetingTool(QtWidgets.QDialog):
                 "Hobbies": 50,
             }
 
+def apply_maya_style(qwidget:QtWidgets.QWidget):
+    """
+    Applies a dark, Maya-inspired color scheme using RGB values.
+    Includes QPushButton, QLineEdit, QTabs, and QGroupBox.
+    """
+    qwidget.setStyleSheet("""
+        QWidget {
+            background-color: rgb(42, 42, 42);  /* Maya dark gray */
+            color: rgb(200, 200, 200);          /* Soft light text */
+        }
+        QPushButton {
+            background-color: rgb(55, 55, 55);
+            color: rgb(210, 210, 210);
+            border: 1px solid rgb(80, 80, 80);
+            padding: 5px;
+            border-radius: 2px;
+        }
+        QPushButton:hover {
+            background-color: rgb(70, 100, 130);
+            border: 1px solid rgb(100, 130, 160);
+        }
+        QLineEdit, QTextEdit, QPlainTextEdit {
+            background-color: rgb(60, 60, 60);
+            color: rgb(230, 230, 230);
+            border: 1px solid rgb(80, 80, 80);
+        }
+        QLabel {
+            color: rgb(200, 200, 200);
+        }
+
+        /* QTabWidget styling */
+        QTabWidget::pane {
+            border: 1px solid rgb(80, 80, 80);
+            background: rgb(42, 42, 42);
+        }
+        QTabBar::tab {
+            background: rgb(55, 55, 55);
+            color: rgb(210, 210, 210);
+            padding: 6px 10px;
+            border: 1px solid rgb(80, 80, 80);
+            border-bottom: none; /* merge with content */
+            min-width: 80px;
+        }
+        QTabBar::tab:selected {
+            background: rgb(70, 100, 130);
+            border-color: rgb(100, 130, 160);
+            color: rgb(240, 240, 240);
+        }
+        QTabBar::tab:hover:!selected {
+            background: rgb(65, 65, 65);
+        }
+        /* QGroupBox with border color + title padding fix */
+        QGroupBox {
+            border: 1px solid rgb(80, 80, 80);
+            margin-top: 12px; /* space for the title text */
+        }
+        QGroupBox::title {
+            subcontrol-origin: margin;
+            subcontrol-position: top left;
+            padding: 0 4px; /* horizontal padding around title */
+        }
+    """)
 
 def launch(open_on_details=False):
     """Launch the budgeting tool window as a standalone PyQt app."""
